scripts/train_complementarity_models.py
"""
训练 3D CNN 和 PointNet++ 模型（Breaking Bad 数据集专用）

使用说明：
1. 确保 Breaking Bad 数据集已准备好
2. 运行此脚本训练模型
3. 训练完成后，权重会自动保存到 pretrained_weights/breaking_bad/ 目录
4. 边界验证模块会自动加载这些预训练权重

使用方法：
    python scripts/train_complementarity_models.py --dataset breaking_bad --model all --epochs 100
"""

import argparse
import torch
import yaml
from pathlib import Path
from datetime import datetime


def train_cnn_3d(config, dataset_root, epochs=100, batch_size=8):
    """
    训练 3D CNN 模型用于形状互补性检查
    
    Args:
        config: 模型配置
        dataset_root: 数据集根目录
        epochs: 训练轮次
        batch_size: 批次大小
    """
    print("=" * 60)
    print("训练 3D CNN 模型 (Breaking Bad 数据集)")
    print("=" * 60)
    
    from src.models.cnn_3d import PointNet3DCNN, Light3DCNN
    
    # 创建模型
    device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
    print(f"\n使用设备：{device}")
    
    use_light_version = config.get('USE_LIGHT_VERSION', False)
    if use_light_version:
        print("使用 3D CNN 轻量版")
        model = Light3DCNN().to(device)
    else:
        print("使用 3D CNN 完整版")
        model = PointNet3DCNN(config.get('MODEL', {})).to(device)
    
    # TODO: 实现训练循环
    # 1. 加载 Breaking Bad 数据集
    # 2. 定义损失函数和优化器
    # 3. 训练模型
    # 4. 保存权重到 pretrained_weights/breaking_bad/cnn3d_breaking_bad_best.pth
    
    print("\n[训练功能待实现]")
    print("建议训练流程：")
    print("1. 使用 Breaking Bad 数据集构建训练样本")
    print("2. 定义互补性判断的损失函数")
    print("3. 训练模型并定期验证")
    print("4. 保存最佳权重到：pretrained_weights/breaking_bad/cnn3d_breaking_bad_best.pth")
    
    # 模拟保存权重（实际训练时需要替换为真实训练代码）
    output_path = Path('pretrained_weights/breaking_bad/cnn3d_breaking_bad_best.pth')
    output_path.parent.mkdir(parents=True, exist_ok=True)
    
    # 保存模型状态
    # torch.save(model.state_dict(), output_path)
    # print(f"\n✓ 模型权重已保存到：{output_path}")
    
    return output_path


def train_pointnet2(config, dataset_root, epochs=100, batch_size=8):
    """
    训练 PointNet++ 模型用于点云特征提取
    
    Args:
        config: 模型配置
        dataset_root: 数据集根目录
        epochs: 训练轮次
        batch_size: 批次大小
    """
    print("=" * 60)
    print("训练 PointNet++ 模型 (Breaking Bad 数据集)")
    print("=" * 60)
    
    from src.models.pointnet2 import PointNet2Encoder, PointNet2SSG
    
    # 创建模型
    device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
    print(f"\n使用设备：{device}")
    
    use_ssg = config.get('USE_SSG', True)
    if use_ssg:
        print("使用 PointNet++ SSG (简化版)")
        model = PointNet2SSG(config).to(device)
    else:
        print("使用 PointNet++ Encoder (完整版)")
        model = PointNet2Encoder(config).to(device)
    
    # TODO: 实现训练循环
    # 1. 加载 Breaking Bad 数据集
    # 2. 定义损失函数和优化器
    # 3. 训练模型
    # 4. 保存权重到 pretrained_weights/breaking_bad/pointnet2_breaking_bad_best.pth
    
    print("\n[训练功能待实现]")
    print("建议训练流程：")
    print("1. 使用 Breaking Bad 数据集构建训练样本")
    print("2. 使用对比学习或监督学习训练特征提取能力")
    print("3. 训练模型并定期验证")
    print("4. 保存最佳权重到：pretrained_weights/breaking_bad/pointnet2_breaking_bad_best.pth")
    
    # 模拟保存权重
    output_path = Path('pretrained_weights/breaking_bad/pointnet2_breaking_bad_best.pth')
    output_path.parent.mkdir(parents=True, exist_ok=True)
    
    # 保存模型状态
    # torch.save(model.state_dict(), output_path)
    # print(f"\n✓ 模型权重已保存到：{output_path}")
    
    return output_path


def main():
    parser = argparse.ArgumentParser(description='训练互补性检查模型')
    parser.add_argument('--dataset', type=str, default='breaking_bad', 
                       help='数据集名称 (default: breaking_bad)')
    parser.add_argument('--model', type=str, default='all', 
                       choices=['cnn', 'pointnet2', 'all'],
                       help='要训练的模型 (default: all)')
    parser.add_argument('--epochs', type=int, default=100,
                       help='训练轮次 (default: 100)')
    parser.add_argument('--batch-size', type=int, default=8,
                       help='批次大小 (default: 8)')
    parser.add_argument('--config', type=str, default='configs/breaking_bad.yaml',
                       help='配置文件路径 (default: configs/breaking_bad.yaml)')
    
    args = parser.parse_args()
    
    # 加载配置
    config_path = Path(args.config)
    if config_path.exists():
        with open(config_path, 'r', encoding='utf-8') as f:
            base_config = yaml.safe_load(f)
    else:
        print(f"⚠ 未找到配置文件 {config_path}，使用默认配置")
        base_config = {}
    
    # 训练指定的模型
    if args.model in ['cnn', 'all']:
        cnn_config = {
            'USE_LIGHT_VERSION': True,
            'MODEL': base_config.get('MODEL', {})
        }
        train_cnn_3d(cnn_config, args.dataset, args.epochs, args.batch_size)
    
    if args.model in ['pointnet2', 'all']:
        pointnet2_config = {
            'INPUT_DIM': 3,
            'OUTPUT_DIM': 256,
            'USE_SSG': True
        }
        train_pointnet2(pointnet2_config, args.dataset, args.epochs, args.batch_size)
    
    print("\n" + "=" * 60)
    print("训练准备完成！")
    print("=" * 60)
    print("\n下一步：")
    print("1. 实现具体的训练循环代码")
    print("2. 准备 Breaking Bad 数据集")
    print("3. 开始训练模型")
    print("4. 权重将自动被边界验证模块使用")


if __name__ == '__main__':
    main()
