requirements.txt
numpy
scipy
open3d
matplotlib
pyyaml
tqdm
scikit-learn
opencv-python
faiss-cpu
trimesh
torch
torchvision